Finish the refactoring for a function cargobull
Remove all kinds of objects inside the middleware, where applicable.
This is mainly for the reason that creating an object that needs
discarding after a dispatch is memory waste and takes up runtime.

The call path is much clearer now. And can be extended too.

Design decisions that still stand: No intention to add callbacks
inside service classes. The service is still a class and that
is unlikely to be changed because extending a module is far from
easy in ruby and sometimes it may be beneficial for development.
That said, no intention is made to support inheritance for services,
as this will lead to very unreadable code. If you need an aggregation,
call a service layer or use the service class as a sub-dispatcher to
your middleware.